updated dependencies, added common directory, added testing for ts files. fixed unhandled promises
