+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@osu-cass/sb-components",
-  "version": "1.4.2",
+  "version": "1.5.0-alpha.1",
   "description": "Shared components for Smarter Balanced",
   "main": "./lib/index.js",
   "types": "./lib/src/index.d.ts",
@@ -50,36 +50,35 @@
   },
   "homepage": "https://github.com/osu-cass/sb-components#readme",
   "devDependencies": {
-    "@osu-cass/tslint-config": "^1.0.8-alpha.3",
-    "@storybook/addon-a11y": "^3.3.13",
-    "@storybook/addon-actions": "^3.3.13",
+    "@osu-cass/tslint-config": "^1.1.0",
+    "@storybook/addon-a11y": "^3.3.15",
+    "@storybook/addon-actions": "^3.3.15",
     "@storybook/addon-console": "^1.0.0",
-    "@storybook/addon-storyshots": "^3.3.13",
-    "@storybook/addons": "^3.3.13",
-    "@storybook/channels": "^3.3.13",
-    "@storybook/cli": "^3.3.13",
-    "@storybook/react": "^3.3.13",
+    "@storybook/addon-storyshots": "^3.3.15",
+    "@storybook/addons": "^3.3.15",
+    "@storybook/channels": "^3.3.15",
+    "@storybook/cli": "^3.3.15",
+    "@storybook/react": "^3.3.15",
     "@types/bootstrap": "3.3.32",
     "@types/enzyme": "^3.1.9",
     "@types/enzyme-adapter-react-16": "^1.0.2",
     "@types/google.analytics": "0.0.29",
-    "@types/jest": "^22.1.3",
+    "@types/jest": "^22.2.0",
     "@types/jquery": "2.0.40",
     "@types/jsdom": "^11.0.4",
-    "@types/node": "^8.0.58",
+    "@types/node": "^9.4.7",
     "@types/react": "^16.0.34",
     "@types/react-dom": "^16.0.3",
     "@types/react-modal": "2.2.2",
     "@types/react-router": "^4.0.12",
     "@types/react-router-dom": "4.0.5",
-    "@types/storybook__addon-actions": "~3.0.2",
+    "@types/storybook__addon-actions": "~3.0.3",
     "@types/storybook__react": "^3.0.7",
     "awesome-typescript-loader": "^3.4.1",
     "copy-webpack-plugin": "^4.3.1",
     "coveralls": "^3.0.0",
     "css-loader": "^0.28.8",
-    "dts-generator": "^2.1.0",
-    "enhanced-resolve": "^4.0.0-beta.4",
+    "enhanced-resolve": "^4.0.0",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
     "enzyme-to-json": "^3.3.1",
@@ -104,7 +103,7 @@
     "react-router-dom": "4.1.1",
     "react-test-renderer": "~16.2.0",
     "style-loader": "^0.19.1",
-    "ts-jest": "~22.0.2",
+    "ts-jest": "~22.4.1",
     "ts-loader": "^3.5.0",
     "tslint": "^5.9.1",
     "tslint-language-service": "^0.9.8",
@@ -144,9 +143,9 @@
     },
     "testRegex": "(stories|src)/.*(test)\\.(js|ts|tsx?)$",
     "collectCoverageFrom": [
-      "src/**/*.tsx",
-      "!src/**/__tests__/*",
-      "!src/index.ts",
+      "src/**/*.ts*",
+      "!src/**/__tests__/**",
+      "!src/*.ts*",
       "!**/node_modules/**"
     ],
     "globals": {
